pci probe

+++ include/kernel/io.h
@@ -3,14 +3,14 @@
 
 #include <stdint.h>
 
-uint8_t inb(uint16_t port);
-void outb(uint16_t port, uint8_t value);
+uint8_t inb(uint32_t port);
+void outb(uint32_t port, uint8_t value);
 
-uint16_t inw(uint16_t port);
-void outw(uint16_t port, uint16_t value);
+uint16_t inw(uint32_t port);
+void outw(uint32_t port, uint16_t value);
 
-uint32_t inl(uint16_t port);
-void outl(uint16_t port, uint32_t value);
+uint32_t inl(uint32_t port);
+void outl(uint32_t port, uint32_t value);
 
 void io_wait();
